frontend/requirements.txt
django
gunicorn
requests
dotenv
pyyaml